BodyPart2
It is an alien in the universe and it starts off with their eyes closed
but if you move your mouse the eyes changes to different colors. Also if
you click the alien then it will change colors and if you press any of
the keys then it will change to a different place in space. Also, the
alien's 3rd eye constantly change it to random colors.